manages to save data to the database
